+++ packages/analytics-js-cookies/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.5.15](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-cookies@0.5.14...@rudderstack/analytics-js-cookies@0.5.15) (2026-09-17)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.30.0`
+
 ## [0.5.14](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-cookies@0.5.13...@rudderstack/analytics-js-cookies@0.5.14) (2026-09-09)
 
 ### Dependency Updates