+++ packages/analytics-js-service-worker/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.3.16](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.3.15...@rudderstack/analytics-js-service-worker@3.3.16) (2026-09-17)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.30.0`
+
 ## [3.3.15](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.3.14...@rudderstack/analytics-js-service-worker@3.3.15) (2026-09-09)
 
 ### Dependency Updates